refactor: replace HelperService with SortService and update component dependencies
