+++ bmusic-ui/src/app/home/artists/artist/artist.component.ts
@@ -7,7 +7,7 @@ import { Song } from '../../../service/models/song.interface';
 import { ApiService } from '../../../service/api.service';
 import { AlbumStateService } from '../../../service/states/album.state.service';
 import { SongsListComponent } from '../../shared/songs-list/songs-list.component';
-import { HelperService } from '../../../service/states/helper.service';
+import { SortService } from '../../../service/states/shared/sort.service';
 
 @Component({
   selector: 'app-artist',
@@ -31,7 +31,7 @@ export class ArtistComponent implements OnInit {
     private artistState: ArtistStateService,
     private apiService: ApiService,
     private albumState: AlbumStateService,
-    private helper: HelperService
+    private helper: SortService
   ) {}
 
   ngOnInit(): void {
