+++ bmusic-ui/src/app/service/states/album.state.service.ts
@@ -8,11 +8,10 @@ import { environment } from '../../../environments/environment';
 })
 export class AlbumStateService {
   private albumsSubject = new BehaviorSubject<Albums[]>([]);
-  public albums$: Observable<Albums[]> = this.albumsSubject.asObservable();
+  public albums$ = this.albumsSubject.asObservable();
 
   private currentAlbumSubject = new BehaviorSubject<CurrentAlbum | null>(null);
-  public currentAlbum$: Observable<CurrentAlbum | null> =
-    this.currentAlbumSubject.asObservable();
+  public currentAlbum$ = this.currentAlbumSubject.asObservable();
 
   setAlbums(albums: Albums[]) {
     this.albumsSubject.next([...albums]);
@@ -32,9 +31,6 @@ export class AlbumStateService {
     }
   }
 
-  getAlbums(): Albums[] {
-    return this.albumsSubject.getValue();
-  }
   getAlbumCover(albumId: number): string {
     const currentAlbums = this.albumsSubject.getValue();
     const currentAlbum = currentAlbums.find((album) => album.id === albumId);
